+++ src/main/java/com/anasdidi/superapp/verticle/helloworld/dto/HelloWorldBaseReqDto.java
@@ -1,5 +1,5 @@
 /* (C) Anas Juwaidi Bin Mohd Jeffry. All rights reserved. */
-package com.anasdidi.superapp.helloworld.dto;
+package com.anasdidi.superapp.verticle.helloworld.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
