+++ src/main/java/com/anasdidi/superapp/verticle/helloworld/dto/HelloWorldBaseResDto.java
@@ -1,4 +1,4 @@
 /* (C) Anas Juwaidi Bin Mohd Jeffry. All rights reserved. */
-package com.anasdidi.superapp.helloworld.dto;
+package com.anasdidi.superapp.verticle.helloworld.dto;
 
 public interface HelloWorldBaseResDto {}
