+++ src/main/java/com/anasdidi/superapp/verticle/helloworld/service/impl/GreetingService.java
@@ -1,12 +1,13 @@
 /* (C) Anas Juwaidi Bin Mohd Jeffry. All rights reserved. */
-package com.anasdidi.superapp.helloworld.service.impl;
+package com.anasdidi.superapp.verticle.helloworld.service.impl;
 
-import com.anasdidi.superapp.helloworld.dto.HelloWorldGreetingReqDto;
-import com.anasdidi.superapp.helloworld.dto.HelloWorldGreetingResDto;
-import com.anasdidi.superapp.helloworld.service.HelloWorldService;
 import io.vertx.core.json.JsonObject;
 import io.vertx.openapi.validation.RequestParameter;
 import java.util.Map;
+
+import com.anasdidi.superapp.verticle.helloworld.dto.HelloWorldGreetingReqDto;
+import com.anasdidi.superapp.verticle.helloworld.dto.HelloWorldGreetingResDto;
+import com.anasdidi.superapp.verticle.helloworld.service.HelloWorldService;
 
 public class GreetingService
     extends HelloWorldService<HelloWorldGreetingReqDto, HelloWorldGreetingResDto> {
